+++ rr/dname.js
@@ -1,4 +1,4 @@
-import net from 'net'
+import net from 'node:net'
 
 import RR from '../rr.js'
 import * as TINYDNS from '../lib/tinydns.js'
@@ -10,10 +10,10 @@ export default class DNAME extends RR {
 
   /****** Resource record specific setters   *******/
   setTarget(val) {
-    if (!val) throw new Error('DNAME: target is required')
+    if (!val) this.throwHelp('DNAME: target is required')
 
     if (net.isIPv4(val) || net.isIPv6(val))
-      throw new Error(`DNAME: target must be a domain name, ${this.citeRFC()}`)
+      this.throwHelp(`DNAME: target must be a domain name`)
 
     this.isFullyQualified('DNAME', 'target', val)
     this.isValidHostname('DNAME', 'target', val)
@@ -42,7 +42,7 @@ export default class DNAME extends RR {
   fromTinydns(opts) {
     // DNAME via generic, :fqdn:n:rdata:ttl:timestamp:lo
     const [fqdn, n, rdata, ttl, ts, loc] = opts.tinyline.substring(1).split(':')
-    if (n != 39) throw new Error('DNAME fromTinydns, invalid n')
+    if (n != 39) this.throwHelp('DNAME fromTinydns, invalid n')
 
     return new DNAME({
       type: 'DNAME',