+++ rr/soa.js
@@ -28,8 +28,7 @@ export default class SOA extends RR {
     // RNAME (email of admin)  (escape . with \)
     this.isValidHostname('SOA', 'RNAME', val)
     this.isFullyQualified('SOA', 'RNAME', val)
-    if (/@/.test(val))
-      throw new Error(`SOA rname replaces @ with a . (dot), ${this.citeRFC()}`)
+    if (/@/.test(val)) this.throwHelp(`SOA rname replaces @ with a . (dot)`)
 
     // RFC 4034: letters in the DNS names are lower cased
     this.set('rname', val.toLowerCase())